# feat: Do not use regex for simple pattern-replacements to improve performance